requirements.txt
python-dotenv==0.20.0
ibm-watson==6.0.0
pyaudio==0.2.11
websocket-client==1.3.3
